First commit for REST.

+++ src/main/java/rest/bibliopole/repository/BookRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import rest.bibliopole.model.Book;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
 
 }
